requirements_updated.txt
# 🏭 КОНТЕНТ-ФАБРИКА 2025 - ПОЛНЫЙ СТЕК ЗАВИСИМОСТЕЙ
# =====================================================

# 🎬 ВИДЕО И МЕДИА ОБРАБОТКА
moviepy==1.0.3              # Основная библиотека обработки видео
ffmpeg-python>=0.2.0        # FFmpeg wrapper  
pillow>=10.0.0              # Обработка изображений
opencv-python>=4.7.0        # Компьютерное зрение
imageio>=2.31.0             # Ввод/вывод изображений
imageio-ffmpeg>=0.4.8       # FFmpeg для imageio

# 🤖 ИСКУССТВЕННЫЙ ИНТЕЛЛЕКТ
torch>=2.0.0                # PyTorch для нейронных сетей
torchvision>=0.15.0         # Компьютерное зрение для PyTorch
transformers>=4.30.0        # Hugging Face трансформеры
openai>=1.0.0               # OpenAI API

# 🌐 ВЕБ И API ИНТЕГРАЦИИ  
aiohttp>=3.8.5              # Асинхронные HTTP запросы
requests>=2.31.0            # HTTP библиотека
httpx>=0.24.1               # Современный HTTP клиент
flask>=2.3.0                # Веб-фреймворк
beautifulsoup4>=4.12.0      # Парсинг HTML

# 📱 СОЦИАЛЬНЫЕ ПЛАТФОРМЫ
google-api-python-client>=2.95.0   # YouTube Data API
google-auth>=2.22.0               # Google Authentication
google-auth-oauthlib>=1.0.0       # OAuth для Google
google-auth-httplib2>=0.1.0       # HTTP lib для Google
yt-dlp>=2023.7.6                  # Скачивание видео
python-telegram-bot>=20.3         # Telegram Bot API

# 📊 АНАЛИЗ ДАННЫХ И СТАТИСТИКА
pandas>=2.0.0               # Анализ данных
numpy>=1.24.3               # Численные вычисления
matplotlib>=3.7.0           # Графики и визуализация
seaborn>=0.12.0             # Статистическая визуализация
scipy>=1.10.0               # Научные вычисления

# ⚡ АСИНХРОННОСТЬ И ПЛАНИРОВАНИЕ
aiofiles>=23.1.0            # Асинхронная работа с файлами
schedule>=1.2.0             # Планировщик задач
celery>=5.3.0               # Очереди задач

# 🗄️ БАЗЫ ДАННЫХ
sqlalchemy>=2.0.0           # ORM для баз данных

# 🎵 АУДИО ОБРАБОТКА  
pydub>=0.25.1               # Обработка аудио
librosa>=0.10.0             # Анализ аудио
soundfile>=0.12.1           # Работа с аудиофайлами
whisper>=1.1.10             # OpenAI Whisper для речи
gtts>=2.3.2                 # Google Text-to-Speech

# 🖥️ GUI И ИНТЕРФЕЙСЫ
PyQt6>=6.5.0                # GUI фреймворк

# 🛠️ СИСТЕМНЫЕ УТИЛИТЫ
psutil>=5.9.5               # Системная информация
python-dotenv>=1.0.0        # Переменные окружения
colorama>=0.4.6             # Цветной вывод в терминале
tqdm>=4.65.0                # Progress bars
click>=8.1.0                # CLI интерфейсы

# 🕒 ВРЕМЯ И ЛОКАЛИЗАЦИЯ
pytz>=2023.3                # Часовые пояса
python-dateutil>=2.8.2      # Парсинг дат

# 📄 КОНФИГУРАЦИИ И ДАННЫЕ
pyyaml>=6.0                 # YAML конфигурации
jsonschema>=4.17.0          # Валидация JSON
pydantic>=2.0.0             # Валидация данных

# 📝 ЛОГИРОВАНИЕ И МОНИТОРИНГ
loguru>=0.7.0               # Продвинутое логирование  
prometheus-client>=0.17.0   # Метрики

# 🔧 ТЕСТИРОВАНИЕ
pytest>=7.4.0              # Фреймворк тестирования
pytest-asyncio>=0.21.0     # Асинхронные тесты

# 🌐 ВЕБ-СКРАПИНГ
selenium>=4.11.0            # Браузерная автоматизация
scrapy>=2.9.0               # Web scraping фреймворк

# 🔐 БЕЗОПАСНОСТЬ
cryptography>=41.0.0        # Криптографические операции
python-magic>=0.4.27       # Определение типов файлов

# 📹 ОБРАБОТКА СУБТИТРОВ
pysrt>=1.1.2                # Работа с SRT субтитрами
webvtt-py>=0.4.6            # Работа с WebVTT

# 🚀 AI СЕРВИСЫ (ОПЦИОНАЛЬНО)
stability-sdk               # Stability AI для изображений
replicate                   # Replicate API для AI моделей

# ⚡ ПРОФИЛИРОВАНИЕ И МОНИТОРИНГ
memory-profiler>=0.61.0     # Профилирование памяти
py-spy>=0.3.14              # Профилирование Python